refactor: re-arranged for qemu development and testing
Configuration
-------------

* moved 'cloudinit' to 'qemu.guest.init_using_cloudinit_image'
* moved 'boot_images' to 'qemu.guest.init_using_image'

With these changes, then the "lookup" of the images files in a map of
"guest-name" => "image-info", is removed. Instead then there are only a
single. Having multiple went against the idea that a config-file
described an environment, e.g with a list of images, then it would
describe a list of environments. Thus this "simplification". This also
means that there are now multiple example configuration-files shipped
with the package.

Configuration: documentation
----------------------------

Instead of the documentation is elaborately described, repeatedled in
every configuration-file, then instead, the configuration-file refers to
which scripts uses the configuration-options, and thus, where one can go
to read documentation on the options.

Building qemu
-------------

With "scripts/build.py" then a multi-system build is introduced, that
is, aarch64 is added to the configure of systems to build.

wrapper.py and scripts/*.py
---------------------------

These are adjusted to the changes in the configuration. Additionally,
then the following are renamed:

* build_x86.py renamed to build.py
  - and extended with aarch64 system

* guest_bootimg.py renamed to guest_init_using_image.py
* guest_cloudinit.py renamed to guest_init_using_cloudinit_image.py

Signed-off-by: Simon A. F. Lund <os@safl.dk>