refactor(scripts): remove the 'guest_start_nvme' script
This was an example of how to use the extend the 'guest_start.py' in
order to do some extended qemu-guest setup. However, it is not properly
maintained, ad instead the xNVMe project has a
'cijoe/xnvme_guest_start_nvme.py' which serves as a much better example.

One extension here is the "init_from_cloudinit_image.py" here, the guest
s started in a different fashion for the purpose of running cloudinit,
this is actively used/maintained and will serve as an example inside the
package itself.

Signed-off-by: Simon A. F. Lund <os@safl.dk>